Add genre model.
The genre model will have a many to many relationship to books. So they each
have an optional array of the other.

In `Book` we also define `@Relationship(inverse: \Genre.books)` for the `genre`
property. It essentially establishes a two-way connection between the data
objects, ensuring consistency and simplifying data management.